add long to day 3 part 2 solution

+++ AoC Day 3/AdventOfCodeDay3Part2.cpp
@@ -49,5 +49,5 @@ int main()
     int d = getTrees(map, numOfRows, numOfCols, 7, 1);
     int e = getTrees(map, numOfRows, numOfCols, 1, 2);
     
-    std::cout << a* b * c * d * e << std::endl;
+    std::cout << long(a* b * c * d * e) << std::endl;
 }
